refactor: consolidate landing components and fix missing links
