+++ client/src/modules/auth/Login.jsx
@@ -5,7 +5,7 @@ import { loginUser } from "../../features/auth/authSlice";
 import Input from "../../shared/components/Input";
 import Button from "../../shared/components/Button";
 import { useToast } from "../../shared/components";
-import Navbar from "../../shared/landing/Navbar";
+import Navbar from "../../modules/landing/components/Navbar";
 import { API_URL } from "../../config/env";
 import { z } from "zod";
 import { useDocumentTitle } from "../../hooks/useDocumentTitle";
