+++ client/src/modules/dashboard/DashboardPage.jsx
@@ -5,7 +5,7 @@ import { LogOut, Menu, User, FileText, Target, Briefcase, Video, BadgeCheck, Spa
 
 import { logout } from "../../features/auth/authSlice";
 import Button from "../../shared/components/Button";
-import Navbar from "../../shared/landing/Navbar";
+import Navbar from "../../modules/landing/components/Navbar";
 import Footer from "../../modules/landing/components/Footer";
 
 import { useDocumentTitle } from "../../hooks/useDocumentTitle";
