+++ client/src/modules/dashboard/pages/CoverLetterHistoryPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../../../modules/landing/components/Footer";
 
 import { getCoverLetterHistory } from "../services/dashboardService";
