+++ client/src/modules/job-matcher/pages/JobMatcherPage.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
 import { useNavigate, Link } from "react-router-dom";
 import { Sparkles, FileUp, AlertCircle, Briefcase, ArrowLeft } from "lucide-react";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../../../modules/landing/components/Footer";
 
 import LoadingState from "../../../shared/components/LoadingState";
