+++ client/src/modules/landing/components/Features.jsx
@@ -10,7 +10,7 @@ import {
   Upload,
   Video,
 } from "lucide-react";
-import Card from "../../../shared/landing/Card";
+import Card from "../../../modules/landing/components/Card";
 const ClassroomPreview = () => (
   <div className="grid h-36 grid-cols-[1fr_0.72fr] gap-3 rounded-lg border border-[var(--border)] bg-[var(--background)] p-3">
     <div className="relative rounded-lg bg-[var(--surface)] p-3">
