+++ client/src/modules/landing/pages/NotFoundPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { AlertCircle, ArrowLeft, Home } from "lucide-react";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 
 const NotFoundPage = () => {
   return (
