+++ client/src/modules/landing/pages/TermsOfServicePage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../components/Footer";
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
