+++ client/src/modules/mock-interview/pages/InterviewHistory.jsx
@@ -12,7 +12,7 @@ import {
   FileJson,
   ArrowLeft,
 } from "lucide-react";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../../../modules/landing/components/Footer";
 
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
