+++ client/src/modules/mock-interview/pages/InterviewSession.test.jsx
@@ -23,7 +23,7 @@ let managerHandlers = {};
 let trackEndedHandler;
 let trackRemoveHandler;
 
-vi.mock("../../../shared/landing/Navbar", () => ({
+vi.mock("../../../modules/landing/components/Navbar", () => ({
   default: () => <nav data-testid="navbar" />,
 }));
