+++ client/src/modules/mock-interview/pages/TutorInterviewsList.jsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { Clock, CheckCircle, Video, ArrowRight, User, ArrowLeft } from "lucide-react";
 import { apiRequest } from "../../../services/apiClient.js";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../../../modules/landing/components/Footer";
 
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
