+++ client/src/modules/mock-interview/pages/__tests__/InterviewHistory.test.jsx
@@ -10,7 +10,7 @@ vi.mock("../../services/interviewService", () => ({
   getHistory: vi.fn(),
 }));
 
-vi.mock("../../../../shared/landing/Navbar", () => ({
+vi.mock("../../../../modules/landing/components/Navbar", () => ({
   default: () => <nav data-testid="navbar">Navbar</nav>,
 }));
