+++ client/src/modules/recruiter-jobs/pages/RecruiterApplicantsPage.jsx
@@ -24,7 +24,7 @@ import {
   LayoutList,
   KanbanSquare
 } from 'lucide-react';
-import Navbar from '../../../shared/landing/Navbar';
+import Navbar from '../../../modules/landing/components/Navbar';
 import Footer from "../../../modules/landing/components/Footer";
 
 import { Button, LoadingState, ErrorState, EmptyState, StatusUpdateModal, StatusTimeline } from '../../../shared/components';
