+++ client/src/modules/recruiter-jobs/pages/__tests__/CreateJobPostingPage.test.jsx
@@ -16,7 +16,7 @@ vi.mock('../../services/jobPostingService', () => ({
 }))
 
 // Mock the components
-vi.mock('../../../../shared/landing/Navbar', () => ({
+vi.mock('../../../../modules/landing/components/Navbar', () => ({
   default: () => <nav data-testid="navbar">Navbar</nav>,
 }))
