+++ client/src/modules/student-jobs/pages/MyApplicationsPage.jsx
@@ -15,7 +15,7 @@ import {
   LayoutGrid,
   List,
 } from "lucide-react";
-import Navbar from "../../../shared/landing/Navbar";
+import Navbar from "../../../modules/landing/components/Navbar";
 import Footer from "../../../modules/landing/components/Footer";
 
 import LoadingState from "../../../shared/components/LoadingState";
