--- client/src/shared/landing/index.js
@@ -1,3 +1,0 @@
-export { default as Navbar } from './Navbar';
-export { default as Button } from './Button';
-export { default as Card } from './Card';
